Implemented UI for jobseeker and jobcreator as well as user interface for testing jobs
Changed the Jobseeker and Jobcreator to their own classes to make it easier for the Test_UI to create many of them. They still work as before just instead of outputing to the console, they are outputed to their UI. Jobcreator and jobseeker can still be run seperately if needed.