Added Erick's write to file functionality
Added the write to file functionality implemented by Erick. Edited so what ever was passed to the 'add to display' functions was also written to the file to consolidate the number of print statements. Also changed the writing functionality to writing directly to the file instead of redirecting std.out. This was done because when multiple JobCreators and JobSeekers are run in same process (ie. Test_UI) all creators and seekers end up writing to the same file.